+++ src/redturtle/agidtheme/controlpanel/interfaces.py
@@ -58,7 +58,8 @@ class IRedturtleAgidthemeSettings(model.Schema):
                 default=u'Header link label'),
         description=_(u'header_link_label_desc',
                       default=u'Label for the link in the header of the site'),
-        required=False
+        required=False,
+        default=u'Governo Italiano'
     )
 
     header_link_url = schema.URI(
@@ -66,7 +67,8 @@ class IRedturtleAgidthemeSettings(model.Schema):
                 default=u'Header link url'),
         description=_(u'header_link_url_desc',
                       default=u'URL of the link in the header'),
-        required=False
+        required=False,
+        default='http://www.governo.it'
     )
 
     header_second_link_label = schema.TextLine(